하나의 파일에 여러개의 클래스가 있어서 분리함
