+++ Nanagong/Nanagong/RegisterLessonView/RegisterLessonViewControllersFacotry.swift
@@ -49,111 +49,7 @@ final class RegisterLessonViewControllersFacotry {
         return RegisterLessonInputFormViewFactory(with: parentViewModel)
     }
     
-    private func makeRegisterLessonViewControllerFactory(parentViewModel: RegisterLessionInformationViewModel) -> RegisterLessonViewControllerFactory {
-        return RegisterLessonViewControllerFactory(appDependancyContainer: appDependency, parentViewModel: parentViewModel)
-    }
-}
-
-final class RegisterLessonInputFormViewFactory {
-    
-    private unowned var parentViewModel: RegisterLessionInformationViewModel
-    
-    init(with parenViewModel: RegisterLessionInformationViewModel) {
-        self.parentViewModel = parenViewModel
-    }
-        
-    func makeInputFormView(with viewMeta: SlothInputFormViewMeta) -> UIView {
-        switch viewMeta.inputFormType {
-        case .lessonName:
-            return makeNameInputFormView(with: viewMeta)
-            
-        case .numberOfLessons:
-            return makeNumberOfLessonsInputFormView(with: viewMeta)
-            
-        case .lessonCategory:
-            return makeCategoryInputFormView(with: viewMeta)
-            
-        case .lessonSite:
-            return makeSiteInputFormView(with: viewMeta)
-            
-        default:
-            return UIView()
-        }
-    }
-    
-    private func makeNameInputFormView(with viewMeta: SlothInputFormViewMeta) -> SlothTextFieldInputFormView {
-        return SlothTextFieldInputFormView(viewModel: makeNameInputFormViewModel(with: viewMeta))
-    }
-    
-    private func makeNumberOfLessonsInputFormView(with viewMeta: SlothInputFormViewMeta) -> SlothTextFieldInputFormView {
-        return SlothTextFieldInputFormView(viewModel: makeNumberOfLessonsInputFormViewModel(with: viewMeta))
-    }
-    
-    private func makeCategoryInputFormView(with viewMeta: SlothInputFormViewMeta) -> SlothSelectBoxInputFormView {
-        return SlothSelectBoxInputFormView(viewModel: makeCategoryInputFormViewModel(with: viewMeta))
-    }
-    
-    private func makeSiteInputFormView(with viewMeta: SlothInputFormViewMeta) -> SlothSelectBoxInputFormView {
-        
-        return SlothSelectBoxInputFormView(viewModel: makeSiteInputFormViewModel(with: viewMeta))
-    }
-        
-    private func makeNameInputFormViewModel(with viewMeta: SlothInputFormViewMeta) -> SlothNameInputFormViewModel {
-        let viewModel = SlothNameInputFormViewModel(viewMeta: viewMeta)
-        parentViewModel.bindWithNameViewState(viewModel.$state.eraseToAnyPublisher())
-        
-        return viewModel
-    }
-    
-    private func makeNumberOfLessonsInputFormViewModel(with viewMeta: SlothInputFormViewMeta) -> SlothNumberOfLessonsInputFormViewModel {
-        let viewModel = SlothNumberOfLessonsInputFormViewModel(viewMeta: viewMeta)
-        parentViewModel.bindWithNumberOfLessonsViewState(viewModel.$state.eraseToAnyPublisher())
-        
-        return viewModel
-    }
-    
-    private func makeCategoryInputFormViewModel(with viewMeta: SlothInputFormViewMeta) -> SlothCategoryInputFormViewModel {
-        let viewModel = SlothCategoryInputFormViewModel(viewMeta: viewMeta, inputSelected: parentViewModel.$selectedCategory.eraseToAnyPublisher())
-        parentViewModel.bindWithCategorySelectViewState(viewModel.$state.eraseToAnyPublisher())
-        
-        return viewModel
-    }
-    
-    private func makeSiteInputFormViewModel(with viewMeta: SlothInputFormViewMeta) -> SlothSiteInputFormViewModel {
-        let viewModel = SlothSiteInputFormViewModel(viewMeta: viewMeta, inputSelected: parentViewModel.$selectedSite.eraseToAnyPublisher())
-        parentViewModel.bindWithSiteSelectViewState(viewModel.$state.eraseToAnyPublisher())
-        
-        return viewModel
-    }
-}
-
-final class RegisterLessonViewControllerFactory {
-    
-    private let appDependancyContainer: SlothAppDependencyContainer
-    private let parentViewModel: RegisterLessionInformationViewModel
-    
-    init(appDependancyContainer: SlothAppDependencyContainer, parentViewModel: RegisterLessionInformationViewModel) {
-        self.appDependancyContainer = appDependancyContainer
-        self.parentViewModel = parentViewModel
-    }
-    
-    func makeSelectCategoryViewController(prevSelected: IdNamePairType?) -> SlothPickerViewController {
-        let pickerViewController = makeSlothPickerViewDependencyContainer().makeSelectCategoryViewController(prevSelected: prevSelected)
-        pickerViewController.modalPresentationStyle = .custom
-        pickerViewController.transitioningDelegate = pickerViewController
-        
-        return pickerViewController
-    }
-    
-    func makeSelectSiteyViewController(prevSelected: IdNamePairType?) -> SlothPickerViewController {
-        let pickerViewController = makeSlothPickerViewDependencyContainer().makeSelectSiteViewController(prevSelected: prevSelected)
-        pickerViewController.modalPresentationStyle = .custom
-        pickerViewController.transitioningDelegate = pickerViewController
-        
-        return pickerViewController
-    }
-    
-    private func makeSlothPickerViewDependencyContainer() -> SlothPickerViewDependencyContainer {
-        return SlothPickerViewDependencyContainer(appDependencyContainer: appDependancyContainer, parentViewModel: parentViewModel)
+    private func makeRegisterLessonViewControllerFactory(parentViewModel: RegisterLessionInformationViewModel) -> RegisterLessonInformationViewPickerFactory {
+        return RegisterLessonInformationViewPickerFactory(appDependancyContainer: appDependency, parentViewModel: parentViewModel)
     }
 }
